Ui LatestArrivals And ProductCategory Implemented.

+++ OnlineShop/DM.Domain/CustomerDiscountAgg/CustomerDiscount.cs
@@ -6,12 +6,12 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
     public class CustomerDiscount : EntityBase
     {
         public long ProductId { get; private set; }
-        public long DiscountRate { get; private set; }
+        public int DiscountRate { get; private set; }
         public DateTime StartDate { get; private set; }
         public DateTime EndDate { get; private set; }
         public string Reason { get; private set; }
 
-        public CustomerDiscount(long productId, long discountRate, DateTime startDate, DateTime endDate, string reason)
+        public CustomerDiscount(long productId, int discountRate, DateTime startDate, DateTime endDate, string reason)
         {
             ProductId = productId;
             DiscountRate = discountRate;
@@ -19,7 +19,8 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
             EndDate = endDate;
             Reason = reason;
         }
-        public void Edit(long productId, long discountRate, DateTime startDate, DateTime endDate, string reason)
+
+        public void Edit(long productId, int discountRate, DateTime startDate, DateTime endDate, string reason)
         {
             ProductId = productId;
             DiscountRate = discountRate;
